+++ src/module/xmas/xmas.module.js
@@ -7,7 +7,7 @@ export function xmas() {
   let lstorage_c_moduleXmastree = localStorage.getItem('tc_c_module_xMasTree')
   const xmassTreeId = 'xmastree'
   const mainHtml = document.getElementsByTagName('main')[0]
-  const configItemDesign = document.getElementById('configItem-design-themes')
+  const configItemDesign = document.getElementById('configItem-designThemes')
   const xmastreeHtml = `<div class="module-xmastree module-timeEEWidget" id="${xmassTreeId}">
     <?xml version="1.0" encoding="UTF-8"?>
     <svg xmlns="http://www.w3.org/2000/svg" version="1.1" viewBox="0 0 1200 1200">